refactor(middleware): consolidate Executor + FsBridge under agent/middleware/
Both abstractions are per-agent host/sandbox pairs consumed only by tools.
Co-locate them in `agent/middleware/`:

- Combine Executor interface + HostExecutor + SandboxExecutor → middleware/executor.ts
- Combine FsBridge interface + HostFs + SandboxFs → middleware/fs.ts
- Delete agent/executor.ts, agent/host-executor.ts, sandbox/executor.ts, sandbox/fs-bridge.ts

AgentRuntime now owns the SandboxExecutor lifecycle:

- AgentRuntimeOptions accepts `sandboxBaseConfig`; constructor builds the executor
- `agentRuntime.shutdown()` cleans up containers
- createAgentTools takes `sandboxExecutor` as an explicit parameter
- Drop module-level singleton + configureToolsLayer / shutdownToolsLayer
- app.ts no longer touches the tools layer for lifecycle

Closes #738
